[FIX] point_of_sale: Fix delay barcode scanner
When we make a request to /hw_proxy/sanner there are a timeout of 7,5 sec
After this timeout we wait for 5 sec before send another request.
And we clean the queue of barcode in IoT Box after 5 sec.

So if a cashier scan a product between this "gap" the product is not added to the POS

With this commit we check if barcode is connected to a IoT Box and not a Posbox
And if true we remove the delay before send another request.

closes odoo/odoo#56969

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>